Autogenerate spec version for modpacks
